Update output format.

+++ usr/lib/python3/dist-packages/linuxmusterCli/typers/check_attic.py
@@ -5,6 +5,8 @@ from rich.console import Console
 from rich.table import Table
 
 from linuxmusterTools.common.attic import check_attic_dir
+from .state import state
+from .format import printf
 
 
 console = Console(emoji=False)
@@ -30,6 +32,8 @@ def check(
     commands = ""
     cmd_count = 0
     to_delete = []
+
+    data = []
     for user,details in result.items():
         if details['start']:
             if details['status'] == "killed":
@@ -44,20 +48,24 @@ def check(
         else:
             status = "No information found"
 
+        data.append([user, status, f"{attic_dir}/{user}"])
         attic.add_row(user, status, f"{attic_dir}/{user}")
 
-    console.print(attic)
-    if cmd_count > 0:
-        console.rule(style="blue")
+    if state.format:
+        printf.format(data)
+    else:
+        console.print(attic)
+        if cmd_count > 0:
+            console.rule(style="blue")
 
-    delete_msg = typer.style(commands, fg=typer.colors.RED, bold=True)
-    if cmd_count == 1:
-        typer.confirm(f"You can delete the unnecessary directory by running:\n\n{delete_msg}\nRun all this command now ?", abort=True)
-        for d in to_delete:
-            shutil.rmtree(d)
-            typer.echo(typer.style(f"{d} deleted!", fg=typer.colors.RED))
-    elif cmd_count > 1:
-        typer.confirm(f"You can delete the unnecessary directories by running:\n\n{delete_msg}\nRun all these commands now ?", abort=True)
-        for d in to_delete:
-            shutil.rmtree(d)
-            typer.echo(typer.style(f"{d} deleted!", fg=typer.colors.RED))
+        delete_msg = typer.style(commands, fg=typer.colors.RED, bold=True)
+        if cmd_count == 1:
+            typer.confirm(f"You can delete the unnecessary directory by running:\n\n{delete_msg}\nRun all this command now ?", abort=True)
+            for d in to_delete:
+                shutil.rmtree(d)
+                typer.echo(typer.style(f"{d} deleted!", fg=typer.colors.RED))
+        elif cmd_count > 1:
+            typer.confirm(f"You can delete the unnecessary directories by running:\n\n{delete_msg}\nRun all these commands now ?", abort=True)
+            for d in to_delete:
+                shutil.rmtree(d)
+                typer.echo(typer.style(f"{d} deleted!", fg=typer.colors.RED))
